+++ native/src/main/java/io/ballerina/stdlib/websocket/actions/websocketconnector/WebSocketSyncConnector.java
@@ -23,6 +23,8 @@ import io.ballerina.stdlib.websocket.WebSocketUtil;
 import io.ballerina.stdlib.websocket.client.listener.SyncClientConnectorListener;
 import io.ballerina.stdlib.websocket.server.WebSocketConnectionInfo;
 
+import java.util.concurrent.atomic.AtomicBoolean;
+
 /**
  * Utilities related to websocket synchronous client connector read actions.
  */
@@ -50,6 +52,7 @@ public class WebSocketSyncConnector {
                 WebSocketConstants.ANNOTATION_ATTR_READ_IDLE_TIMEOUT, 0);
         connectionInfo.getWebSocketConnection().addReadIdleStateHandler(readTimeoutInSeconds);
         connectorListener.setCallback(callback);
+        connectorListener.setFutureCompleted(new AtomicBoolean(false));
         connectionInfo.getWebSocketConnection().readNextFrame();
     }